+++ avldb/src/main/scala/scorex/db/LDBFactory.scala
@@ -112,7 +112,6 @@ object LDBFactory extends ScorexLogging {
   private val javaFactory = "org.iq80.leveldb.impl.Iq80DBFactory"
   private val memoryPoolSize = 512 * 1024
 
-
   def setLevelDBParams(factory: DBFactory): Unit = {
     val pushMemoryPool = factory.getClass.getDeclaredMethod("pushMemoryPool", classOf[Int])
     pushMemoryPool.invoke(null, Integer.valueOf(memoryPoolSize))
