+++ src/main/scala/org/ergoplatform/nodeView/mempool/FeeHistogramBin.scala
@@ -1,0 +1,15 @@
+package org.ergoplatform.nodeView.mempool
+
+import io.circe.{Encoder, Json}
+import io.circe.syntax._
+
+case class FeeHistogramBin(nTxns: Int, totalFee: Long)
+
+object FeeHistogramBin {
+
+  implicit val encodeHistogramBin: Encoder[FeeHistogramBin] = (bin: FeeHistogramBin) => Json.obj(
+    ("nTxns", bin.nTxns.asJson),
+    ("totalFee", bin.totalFee.asJson)
+  )
+
+}
